% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_core.R
\name{bayesSYNC_multi_core}
\alias{bayesSYNC_multi_core}
\title{Core multi-study CAVI algorithm with annealing}
\usage{
bayesSYNC_multi_core(
  S,
  n_s,
  p,
  d,
  L_f,
  L_s,
  M_f,
  M_s,
  K,
  C,
  C_g,
  Y,
  Z,
  var_names,
  mean_mean_across_subjects,
  sd_mean_across_subjects,
  anneal,
  list_hyper,
  time_obs,
  bool_var_spec_prob,
  n_g,
  time_g,
  tol_abs,
  tol_rel,
  maxit,
  convergence_rule,
  convergence_rule_requested,
  elbo_role,
  lambda_orth,
  practical_control,
  initialization,
  initialization_control,
  fit_seed,
  n_cpus,
  verbose,
  scale_trace_control = NULL
)
}
\description{
Core multi-study CAVI algorithm with annealing
}
\keyword{internal}
