% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_core.R
\name{compute_elbo_multi}
\alias{compute_elbo_multi}
\title{Compute the complete ordinary multi-study ELBO}
\usage{
compute_elbo_multi(
  Y,
  C,
  list_cp_C,
  total_obs_sj = NULL,
  mu_q_nu_mu,
  Sigma_q_nu_mu,
  mu_q_nu_beta,
  Sigma_q_nu_beta,
  Z,
  mu_q_zeta,
  Sigma_q_zeta,
  mu_q_nu_phi,
  Sigma_q_nu_phi,
  mu_q_xi,
  Sigma_q_xi,
  mu_q_nu_psi,
  Sigma_q_nu_psi,
  mu_q_a,
  mu_q_normal_a,
  Sigma_q_normal_a,
  mu_q_gamma_a,
  mu_q_b_specific,
  mu_q_normal_b,
  Sigma_q_normal_b,
  mu_q_gamma_b,
  mu_q_recip_sigsq_eps,
  mu_q_recip_a_eps,
  mu_q_recip_sigsq_mu,
  mu_q_recip_a_mu,
  mu_q_recip_sigsq_beta,
  mu_q_recip_a_beta,
  mu_q_recip_sigsq_phi,
  mu_q_recip_a_phi,
  mu_q_recip_sigsq_psi,
  mu_q_recip_a_psi,
  mu_q_log_sigsq_eps,
  mu_q_log_sigsq_mu,
  mu_q_log_sigsq_beta,
  mu_q_log_sigsq_phi,
  mu_q_log_sigsq_psi,
  kappa_q_sigsq_eps,
  kappa_q_sigsq_mu,
  kappa_q_sigsq_beta,
  kappa_q_sigsq_phi,
  kappa_q_sigsq_psi,
  lambda_q_sigsq_eps,
  lambda_q_sigsq_mu,
  lambda_q_sigsq_beta,
  lambda_q_sigsq_phi,
  lambda_q_sigsq_psi,
  kappa_q_a,
  mu_q_log_omega_a,
  mu_q_log_1_omega_a,
  mu_q_log_omega_b,
  mu_q_log_1_omega_b,
  c_1_omega_a,
  d_1_omega_a,
  c_1_omega_b,
  d_1_omega_b,
  c_0,
  d_0,
  inv_Sigma_beta,
  S,
  n_s,
  p,
  d,
  L_f,
  L_s,
  M_f,
  M_s,
  K,
  c_val = 1,
  term_a = NULL,
  term_b_specific = NULL,
  expected_rss = NULL,
  expected_rss_sum = NULL,
  lambda_q_a_eps = NULL,
  lambda_q_a_mu = NULL,
  lambda_q_a_beta = NULL,
  lambda_q_a_phi = NULL,
  lambda_q_a_psi = NULL,
  A = 1e+05,
  bool_var_spec_prob = FALSE,
  elbo_history = numeric(),
  monotonicity_tolerance = sqrt(.Machine$double.eps)
)
}
\description{
The returned objective is the T = 1 ELBO of the probability model. During
annealing the core routine deliberately does not call this function.
}
\keyword{internal}
