% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variance.R
\name{compute_rss_cache}
\alias{compute_rss_cache}
\title{Compute and cache all current expected residual sums of squares}
\usage{
compute_rss_cache(
  Y,
  C,
  list_cp_C,
  mu_q_nu_mu,
  Sigma_q_nu_mu,
  mu_q_nu_beta,
  Sigma_q_nu_beta,
  Z,
  mu_q_zeta,
  Sigma_q_zeta,
  mu_q_nu_phi,
  Sigma_q_nu_phi,
  mu_q_xi,
  Sigma_q_xi,
  mu_q_nu_psi,
  Sigma_q_nu_psi,
  mu_q_a,
  term_a,
  mu_q_b_specific,
  term_b_specific,
  S,
  n_s,
  p,
  L_f,
  L_s,
  return_fitted = FALSE
)
}
\description{
Compute and cache all current expected residual sums of squares
}
\keyword{internal}
