% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_structured.R
\name{construct_orthonormal_eigenfunctions}
\alias{construct_orthonormal_eigenfunctions}
\title{Construct L2-orthonormal B-spline basis functions on [0,1]}
\usage{
construct_orthonormal_eigenfunctions(M, n_g = 200, degree = 3)
}
\arguments{
\item{M}{Number of basis functions}

\item{n_g}{Dense grid size}

\item{degree}{B-spline degree (default 3 = cubic)}
}
\value{
List: phi_matrix (n_g x M), t_grid (n_g x 1)
}
\description{
Construct L2-orthonormal B-spline basis functions on [0,1]
}
\keyword{internal}
