% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_structured.R
\name{eval_eigenfun}
\alias{eval_eigenfun}
\title{Evaluate an eigenfunction at arbitrary time points by interpolation}
\usage{
eval_eigenfun(t_pts, phi_dense, t_grid)
}
\arguments{
\item{t_pts}{Time points to evaluate}

\item{phi_dense}{Dense-grid eigenfunction vector}

\item{t_grid}{Dense grid}
}
\value{
Vector of interpolated values
}
\description{
Evaluate an eigenfunction at arbitrary time points by interpolation
}
\keyword{internal}
