% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data_osullivan.R
\name{project_to_osullivan}
\alias{project_to_osullivan}
\title{Project a function evaluated on a grid onto O'Sullivan spline basis}
\usage{
project_to_osullivan(f_vals, K, t_grid, ridge = 1e-06, int_knots = NULL)
}
\arguments{
\item{f_vals}{Function values on t_grid (length n_g)}

\item{K}{Number of O'Sullivan internal knots}

\item{t_grid}{Dense grid of time points (length n_g)}

\item{ridge}{Ridge regularization for projection matrix (default 1e-6)}
}
\value{
Coefficient vector of length K+2
}
\description{
Project a function evaluated on a grid onto O'Sullivan spline basis
}
\keyword{internal}
