% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variance.R
\name{update_a_beta}
\alias{update_a_beta}
\title{Update auxiliary variable for regression coefficient variance}
\usage{
update_a_beta(mu_q_recip_sigsq_beta, A = 1e+05, c_val = 1)
}
\arguments{
\item{mu_q_recip_sigsq_beta}{p x d matrix, E_q[1/sigma^2_beta]}

\item{A}{Half-Cauchy scale}

\item{c_val}{Temperature constant}
}
\value{
List with updated parameters
}
\description{
Update auxiliary variable for regression coefficient variance
}
