% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variance.R
\name{update_a_eps}
\alias{update_a_eps}
\title{Update auxiliary variable for measurement error variance}
\usage{
update_a_eps(mu_q_recip_sigsq_eps, A = 1e+05, c_val = 1)
}
\arguments{
\item{mu_q_recip_sigsq_eps}{S x p matrix, E_q[1/sigma^2_eps]}

\item{A}{Half-Cauchy scale hyperparameter (default 1e5)}

\item{c_val}{Temperature constant c = 1/T}
}
\value{
List with updated mu_q_recip_a_eps, kappa_q_a, lambda_q_a_eps
}
\description{
Update auxiliary variable for measurement error variance
}
