% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variance.R
\name{update_a_mu}
\alias{update_a_mu}
\title{Update auxiliary variable for mean function variance}
\usage{
update_a_mu(mu_q_recip_sigsq_mu, A = 1e+05, c_val = 1)
}
\arguments{
\item{mu_q_recip_sigsq_mu}{S x p matrix, E_q[1/sigma^2_mu]}

\item{A}{Half-Cauchy scale}

\item{c_val}{Temperature constant}
}
\value{
List with updated parameters
}
\description{
Update auxiliary variable for mean function variance
}
