% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variance.R
\name{update_a_phi}
\alias{update_a_phi}
\title{Update auxiliary variable for shared eigenfunction variance}
\usage{
update_a_phi(mu_q_recip_sigsq_phi, A = 1e+05, c_val = 1)
}
\arguments{
\item{mu_q_recip_sigsq_phi}{List of vectors, E_q[1/sigma^2_phi], indexed [[l]][m]}

\item{A}{Half-Cauchy scale}

\item{c_val}{Temperature constant}
}
\value{
List with updated parameters (mu_q_recip_a_phi returned as list of vectors)
}
\description{
Update auxiliary variable for shared eigenfunction variance
}
