% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_variance.R
\name{update_a_psi}
\alias{update_a_psi}
\title{Update auxiliary variable for specific eigenfunction variance}
\usage{
update_a_psi(mu_q_recip_sigsq_psi, A = 1e+05, c_val = 1)
}
\arguments{
\item{mu_q_recip_sigsq_psi}{List of list of vectors, E_q[1/sigma^2_psi], indexed [[s]][[l]][m]}

\item{A}{Half-Cauchy scale}

\item{c_val}{Temperature constant}
}
\value{
List with updated parameters (mu_q_recip_a_psi returned as list of list of vectors)
}
\description{
Update auxiliary variable for specific eigenfunction variance
}
