% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_omega.R
\name{update_omega}
\alias{update_omega}
\title{Update all omega parameters (shared + specific)}
\usage{
update_omega(
  mu_q_gamma_a,
  mu_q_gamma_b = NULL,
  p,
  S,
  c_0 = 1,
  d_0 = NULL,
  c_val = 1,
  bool_var_spec_prob = FALSE
)
}
\arguments{
\item{mu_q_gamma_a}{p x L_f matrix of shared PPIs}

\item{mu_q_gamma_b}{List of S matrices, specific PPIs}

\item{p}{Number of variables}

\item{S}{Number of studies}

\item{c_0}{Beta prior shape1}

\item{d_0}{Beta prior shape2}

\item{c_val}{Temperature constant}

\item{bool_var_spec_prob}{Variable-specific probabilities flag}
}
\value{
List with all updated omega parameters
}
\description{
Update all omega parameters (shared + specific)
}
